Change VAT rate type from int to float across the system
remp/crm#3277

+++ src/Models/PaymentItem/SubscriptionTypePaymentItem.php
@@ -16,7 +16,7 @@ final class SubscriptionTypePaymentItem implements PaymentItemInterface
         private int $subscriptionTypeId,
         string $name,
         float $price,
-        int $vat,
+        float $vat,
         int $count = 1,
         array $meta = [],
         private ?int $subscriptionTypeItemId = null,
